adding tag functionality
