+++ views/feed.erb
@@ -31,6 +31,7 @@
 					<input type="submit" value="Peep!"> </form>
 				<% end %>
 		</div>
+		
 		<% @peeps.each do |peep|%>
 			<article class="container__peep">
 				<div class="container__peep_info">
@@ -41,11 +42,20 @@
 				</div>
 				<div> <img class="avatar" src="/img/avatar1.jpg" alt="avatar">
 					<p class="peep__text">
-						<%= peep.content %>
+
+						<% if !extract_tagged_users(peep.content).empty? %>
+							<% extract_tagged_users(peep.content).each do |username| %>
+								<% peep.content.gsub! "@#{username}", "<a href='/users/#{username}'>@#{username}</a>" %>
+							<% end %>
+							<%= peep.content %>
+						<% else %>
+							<%= peep.content %>
+						<% end %>
+							
 					</p>
 				</div>
 			</article>
-			<% end %>
+		<% end %>
 </body>
 
 </html>
